+++ app/src/main/java/org/simple/clinic/patient/Patient.kt
@@ -490,12 +490,14 @@ data class Patient(
 
     @Query("""
       UPDATE Patient
-      SET eligibleForReassignment = :eligibleForReassignment
+      SET eligibleForReassignment = :eligibleForReassignment, syncStatus = :syncStatus, updatedAt = :updatedAt
       WHERE uuid = :patientUuid
     """)
     abstract fun updatePatientReassignmentEligibilityStatus(
         patientUuid: UUID,
         eligibleForReassignment: Answer,
+        syncStatus: SyncStatus,
+        updatedAt: Instant,
     )
   }
 }