+++ app/src/main/java/org/simple/clinic/patient/PatientRepository.kt
@@ -838,7 +838,9 @@ class PatientRepository @Inject constructor(
   ) {
     database.patientDao().updatePatientReassignmentEligibilityStatus(
         patientUuid = patientUuid,
-        eligibleForReassignment = eligibleForReassignment
+        eligibleForReassignment = eligibleForReassignment,
+        syncStatus = PENDING,
+        updatedAt = Instant.now(utcClock),
     )
   }